+++ react-components/src/components/FundingCredit/FundingCreditTotalExportCost/FundingCreditTotalExportCost.jsx
@@ -1,13 +1,11 @@
 import React, { memo, useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import Services from '@src/Services'
-import { config } from '@src/config'
 import { Input } from '@src/components/Form/Input'
 import { useDebounce } from '@src/components/hooks/useDebounce'
 
 export const FundingCreditTotalExportCost = memo(({ ...data }) => {
   const { estimated_costs_per_unit, formData, currency } = data
-  const { urlCostsAndPricing } = config
   const [formValue, setFormValue] = useState(
     formData['override_estimated_total_cost']
   )
@@ -40,7 +38,7 @@ export const FundingCreditTotalExportCost = memo(({ ...data }) => {
       estimated_costs_per_unit !== 0
         ? {
             buttonTitle: 'Estimate',
-            header: `Your estimate total export cost is GBP ${estimated_costs_per_unit}`,
+            header: `Your estimate total export cost is GBP ${formValue}`,
             content: `<p>
         We calculated this by:
       </p>
@@ -55,23 +53,7 @@ export const FundingCreditTotalExportCost = memo(({ ...data }) => {
         : {},
   }
 
-  return (
-    <>
-      {estimated_costs_per_unit !== 0 ? (
-        <p>
-          To help you, we've created an estimate for you based on the figures
-          you gave in on the Cost and pricing page.
-        </p>
-      ) : (
-        <p>
-          To get an estimate of your total export cost, complete the{' '}
-          <a href={urlCostsAndPricing}>Costs and Pricing</a> section of your
-          Export Plan. Once you're done, you'll see your estimate here.
-        </p>
-      )}
-      <Input {...inputData} />
-    </>
-  )
+  return <Input {...inputData} />
 })
 
 FundingCreditTotalExportCost.propTypes = {